Move youtube functions back into content resolver
