+++ services/graphql-server/src/graphql/resolvers/platform/content.js
@@ -19,7 +19,6 @@ const relatedContent = require('../../utils/related-content');
 const inquiryEmails = require('../../utils/inquiry-emails');
 const connectionProjection = require('../../utils/connection-projection');
 const getDescendantIds = require('../../utils/website-section-child-ids');
-const retrieveYoutubeVideos = require('../../utils/retrieve-youtube-videos');
 const {
   createTitle,
   createDescription,
@@ -27,6 +26,25 @@ const {
   getDefaultContentTypes,
 } = require('../../utils/content');
 const contentTeaser = require('../../utils/content-teaser');
+const googleDataApiClient = require('../../../google-data-api-client');
+
+const retrieveYoutubePlaylistId = async ({ youtube }) => {
+  const playlistId = get(youtube, 'playlistId');
+  if (playlistId) return playlistId;
+
+  const id = get(youtube, 'channelId');
+  const forUsername = get(youtube, 'username');
+  if (id || forUsername) {
+    const payload = {
+      part: 'contentDetails',
+      ...(id && { id }),
+      ...(forUsername && { forUsername }),
+    };
+    const response = await googleDataApiClient.request('youtube.channelList', payload);
+    return get(response, 'items.0.contentDetails.relatedPlaylists.uploads');
+  }
+  return undefined;
+};
 
 const { isArray } = Array;
 
@@ -492,7 +510,13 @@ module.exports = {
    */
   ContentCompany: {
     youtube: ({ youtube = {} }) => youtube,
-    youtubeVideos: retrieveYoutubeVideos,
+    youtubeVideos: async (content, { input }, { basedb }) => {
+      const { limit } = input;
+      const playlistId = await retrieveYoutubePlaylistId(content, basedb);
+      if (!playlistId) return { pageInfo: {}, items: [] };
+      const response = await googleDataApiClient.request('youtube.playlistItems', { playlistId, maxResults: limit });
+      return response;
+    },
   },
 
   /**
